+++ packages/pwa/app/components/item-variant/item-attributes.jsx
@@ -13,7 +13,7 @@ import useBasket from '../../commerce-api/hooks/useBasket'
 import {Flex, Stack, Text} from '@chakra-ui/react'
 import {useItemVariant} from './'
 import PromoPopover from '../promo-popover'
-import {DEFAULT_CURRENCY} from '../../constants'
+import {useCurrency} from '../../hooks'
 
 /**
  * In the context of a cart product item variant, this component renders a styled
@@ -23,6 +23,7 @@ const ItemAttributes = ({includeQuantity, currency, ...props}) => {
     const variant = useItemVariant()
     const basket = useBasket()
     const [promos, setPromos] = useState([])
+    const {currency: activeCurrency} = useCurrency()
 
     // Create a mapping of variation values to their associated attributes. This allows us
     // the render the readable names/labels rather than variation value IDs.
@@ -82,7 +83,7 @@ const ItemAttributes = ({includeQuantity, currency, ...props}) => {
                         <Text as="span" color="green.500">
                             <FormattedNumber
                                 style="currency"
-                                currency={currency || basket.currency || DEFAULT_CURRENCY}
+                                currency={currency || basket.currency || activeCurrency}
                                 value={variant.priceAdjustments[0].price}
                             />
                         </Text>