+++ packages/pwa/app/partials/product-view/index.jsx
@@ -32,7 +32,7 @@ import ImageGallery from '../../components/image-gallery'
 import Breadcrumb from '../../components/breadcrumb'
 import Link from '../../components/link'
 import withRegistration from '../../hoc/with-registration'
-import {DEFAULT_CURRENCY} from '../../constants'
+import {useCurrency} from '../../hooks'
 import {Skeleton as ImageGallerySkeleton} from '../../components/image-gallery'
 import AddToCartModal from '../../components/add-to-cart-modal'
 import RecommendedProducts from '../../components/recommended-products'
@@ -41,6 +41,7 @@ import QuantityPicker from '../../components/quantity-picker'
 
 const ProductViewHeader = ({name, price, currency, category}) => {
     const intl = useIntl()
+    const {currency: activeCurrency} = useCurrency()
     return (
         <VStack mr={4} spacing={2} align="flex-start" marginBottom={[4, 4, 4, 0, 0]}>
             {category && (
@@ -59,7 +60,7 @@ const ProductViewHeader = ({name, price, currency, category}) => {
                 <Text fontWeight="bold" fontSize="md" aria-label="price">
                     {intl.formatNumber(price, {
                         style: 'currency',
-                        currency: currency || DEFAULT_CURRENCY
+                        currency: currency || activeCurrency
                     })}
                 </Text>
             </Skeleton>