+++ packages/pwa/app/utils/url.test.js
@@ -10,10 +10,13 @@ import {
     categoryUrlBuilder,
     productUrlBuilder,
     searchUrlBuilder,
+    getUrlWithLocale,
     homeUrlBuilder,
     rebuildPathWithParams,
     removeQueryParamsFromPath
 } from './url'
+
+import {DEFAULT_LOCALE} from '../constants'
 
 describe('buildUrlSet returns the expected set of urls', () => {
     test('when no values are passed in', () => {
@@ -70,6 +73,17 @@ describe('buildUrlSet returns the expected set of urls', () => {
 })
 
 describe('url builder test', () => {
+    // Save the original `window.location` object to not affect other test
+    const originalLocation = window.location
+
+    beforeEach(() => {
+        delete window.location
+        window.location = {...originalLocation, assign: jest.fn()}
+    })
+    afterEach(() => {
+        // Restore `window.location` to the `jsdom` `Location` object
+        window.location = originalLocation
+    })
     test('searchUrlBuilder returns expect', () => {
         const url = searchUrlBuilder('term')
 
@@ -91,8 +105,68 @@ describe('url builder test', () => {
         const url = homeUrlBuilder('/', 'fr-FR')
         expect(url).toEqual(`/fr-FR/`)
 
-        const homeUrlDefaultLocale = homeUrlBuilder('/', 'en-GB')
+        const homeUrlDefaultLocale = homeUrlBuilder('/', DEFAULT_LOCALE)
         expect(homeUrlDefaultLocale).toEqual(`/`)
+    })
+
+    test('getUrlWithLocale returns expected for PLP', () => {
+        const location = new URL('http://localhost:3000/it-IT/category/newarrivals-womens')
+
+        window.location = location
+
+        const relativeUrl = getUrlWithLocale('fr-FR')
+        expect(relativeUrl).toEqual(`/fr-FR/category/newarrivals-womens`)
+    })
+
+    test('getUrlWithLocale returns expected for PLP without refine param', () => {
+        const location = new URL(
+            'http://localhost:3000/it-IT/category/newarrivals-womens?limit=25&refine=c_refinementColor%3DBianco&sort=best-matches&offset=25'
+        )
+
+        window.location = location
+
+        const relativeUrl = getUrlWithLocale('fr-FR', {
+            disallowParams: ['refine']
+        })
+        expect(relativeUrl).toEqual(
+            `/fr-FR/category/newarrivals-womens?limit=25&sort=best-matches&offset=25`
+        )
+    })
+
+    test('getUrlWithLocale returns expected for PLP', () => {
+        const location = new URL('http://localhost:3000/it-IT/category/newarrivals-womens')
+
+        window.location = location
+
+        const relativeUrl = getUrlWithLocale('fr-FR')
+        expect(relativeUrl).toEqual(`/fr-FR/category/newarrivals-womens`)
+    })
+
+    test('getUrlWithLocale returns expected for Homepage', () => {
+        const location = new URL('http://localhost:3000/it-IT/')
+
+        window.location = location
+
+        const relativeUrl = getUrlWithLocale('fr-FR')
+        expect(relativeUrl).toEqual(`/fr-FR`)
+    })
+
+    test('getUrlWithLocale returns expected for Homepage with Default locale', () => {
+        const location = new URL('http://localhost:3000/it-IT/')
+
+        window.location = location
+
+        const relativeUrl = getUrlWithLocale(DEFAULT_LOCALE)
+        expect(relativeUrl).toEqual(`/`)
+    })
+
+    test('getUrlWithLocale returns expected for Homepage without trailing slash', () => {
+        const location = new URL('http://localhost:3000/it-IT')
+
+        window.location = location
+
+        const relativeUrl = getUrlWithLocale(DEFAULT_LOCALE)
+        expect(relativeUrl).toEqual(`/`)
     })
 })